+++ program/rust/src/accounts.rs
@@ -36,6 +36,11 @@ mod permission;
 mod price;
 mod product;
 
+#[cfg(test)]
+pub use product::{
+    account_has_key_values,
+    create_pc_str_t,
+};
 pub use {
     mapping::MappingAccount,
     permission::{
