+++ program/rust/src/accounts/product.rs
@@ -100,3 +100,49 @@ pub fn read_pc_str_t(source: &[u8]) -> Result<&[u8], ProgramError> {
         }
     }
 }
+
+#[cfg(test)]
+pub fn create_pc_str_t(s: &str) -> Vec<u8> {
+    let mut v = vec![s.len() as u8];
+    v.extend_from_slice(s.as_bytes());
+    v
+}
+
+// Check that the key-value list in product_account equals the strings in expected
+// Returns an Err if the account data is incorrectly formatted and the comparison cannot be
+// performed.
+#[cfg(test)]
+pub fn account_has_key_values(
+    product_account: &AccountInfo,
+    expected: &[&str],
+) -> Result<bool, ProgramError> {
+    let account_size: usize = try_convert(
+        load_checked::<ProductAccount>(product_account, crate::c_oracle_header::PC_VERSION)?
+            .header
+            .size,
+    )?;
+    let mut all_account_data = product_account.try_borrow_mut_data()?;
+    let kv_data = &mut all_account_data[size_of::<ProductAccount>()..account_size];
+    let mut kv_idx = 0;
+    let mut expected_idx = 0;
+
+    while kv_idx < kv_data.len() {
+        let key = read_pc_str_t(&kv_data[kv_idx..])?;
+        if key[0] != try_convert::<_, u8>(key.len())? - 1 {
+            return Ok(false);
+        }
+
+        if &key[1..] != expected[expected_idx].as_bytes() {
+            return Ok(false);
+        }
+
+        kv_idx += key.len();
+        expected_idx += 1;
+    }
+
+    if expected_idx != expected.len() {
+        return Ok(false);
+    }
+
+    Ok(true)
+}
