+++ program/rust/src/lib.rs
@@ -1,3 +1,4 @@
+#![deny(warnings)]
 // Allow non upper case globals from C
 #![allow(non_upper_case_globals)]
